requirements.txt
torch>=1.10.0
torchvision>=0.11.0
torchaudio>=0.10.0
transformers>=4.20.0
pytube>=12.1.0
Pillow>=9.0.0
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=0.24.2
matplotlib>=3.4.0
tqdm>=4.62.0
requests>=2.26.0
python-dotenv>=0.19.0
pytest>=6.2.5
tensorboard>=2.8.0
wandb>=0.12.0
accelerate>=0.12.0 